+++ BitPayLib/class-bitpayipnprocess.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace BitPayLib;
 
+use BitPayLib\Exception\BitPayInvalidOrder;
 use BitPaySDK\Model\Facade;
 use BitPaySDK\Model\Invoice\Invoice;
 use WC_Order;
@@ -20,6 +21,8 @@ use WP_REST_Request;
 class BitPayIpnProcess {
 
 	use WpDbHelper;
+
+	private const FINAL_WC_ORDER_STATUSES = array( 'wc-refunded', 'wc-cancelled', 'wc-failed' );
 
 	private BitPayCheckoutTransactions $bitpay_checkout_transactions;
 	private BitPayLogger $logger;
@@ -39,43 +42,64 @@ class BitPayIpnProcess {
 	public function execute( WP_REST_Request $request ): void {
 		$data = $request->get_body();
 
-		$data       = json_decode( $data, true, 512, JSON_THROW_ON_ERROR );
-		$event      = $data['event'] ?? null;
-		$data       = $data['data'] ?? null;
-		$invoice_id = $data['id'] ?? null;
+		$data                = json_decode( $data, true, 512, JSON_THROW_ON_ERROR );
+		$event               = $data['event'] ?? null;
+		$data                = $data['data'] ?? null;
+		$data['event']       = $event;
+		$data['requestDate'] = date( 'Y-m-d H:i:s' );
+		$invoice_id          = $data['id'] ?? null;
 
 		$this->logger->execute( $data, 'INCOMING IPN', true );
 		if ( ! $event || ! $data || ! $invoice_id ) {
 			$this->logger->execute( 'Wrong IPN request', 'INCOMING IPN ERROR', false, true );
+			return;
 		}
 
 		try {
 			$bitpay_invoice = $this->factory->create()->getInvoice( $invoice_id, Facade::POS, false );
-			$order          = new WC_Order( $bitpay_invoice->getOrderId() );
+			do_action( 'bitpay_checkout_woocoomerce_after_get_invoice', $bitpay_invoice );
+			$order = new WC_Order( $bitpay_invoice->getOrderId() );
 			$this->validate_order( $order, $invoice_id );
 			$this->process( $bitpay_invoice, $order, $event['name'] );
+		} catch ( BitPayInvalidOrder $e ) {
+			return; // do nothing.
 		} catch ( \Exception $e ) {
 			$this->logger->execute( $e->getMessage(), 'INCOMING IPN ERROR', false, true );
 		}
 	}
 
 	private function validate_order( WC_Order $order, string $invoice_id ): void {
+		do_action( 'bitpay_checkout_woocoomerce_validate_wc_order', $order, $invoice_id );
+
 		if ( $order->get_payment_method() !== 'bitpay_checkout_gateway' ) {
 			$message = 'Order id = ' . $order->get_id() . ', BitPay invoice id = ' . $invoice_id
 				. '. Current payment method = ' . $order->get_payment_method();
 			$this->logger->execute( $message, 'Ignore IPN', true );
-			die();
+			throw new BitPayInvalidOrder();
 		}
 
 		if ( $this->bitpay_checkout_transactions->count_transaction_id( $invoice_id ) !== 1 ) {
 			$message = 'Order id = ' . $order->get_id() . ', BitPay invoice id = ' . $invoice_id
 				. '. Wrong transaction id ' . $invoice_id;
 			$this->logger->execute( $message, 'Ignore IPN', true );
-			die();
+			throw new BitPayInvalidOrder();
 		}
 	}
 
 	private function process( Invoice $bitpay_invoice, WC_Order $order, string $event_name ): void {
+		do_action( 'bitpay_checkout_woocoomerce_before_process', $bitpay_invoice, $order, $event_name );
+
+		// do not process order when it is in a final status.
+		if ( $this->has_final_status( $order ) ) {
+			$message = 'Order ' . $order->get_id() . ' has final status ' . $order->get_status()
+				. '. Order status from ' . $event_name . ' was not updated for event ' . $event_name;
+			$order->add_order_note( $message );
+
+			return;
+		}
+
+		do_action( 'bitpay_checkout_woocoomerce_process', $bitpay_invoice, $order, $event_name );
+
 		switch ( $event_name ) {
 			case 'invoice_completed':
 				$this->process_completed( $bitpay_invoice, $order );
@@ -99,6 +123,8 @@ class BitPayIpnProcess {
 				$this->process_refunded( $bitpay_invoice, $order );
 				break;
 		}
+
+		do_action( 'bitpay_checkout_woocoomerce_after_process', $bitpay_invoice, $order, $event_name );
 
 		$this->bitpay_checkout_transactions->update_transaction_status( $bitpay_invoice );
 
@@ -146,12 +172,12 @@ class BitPayIpnProcess {
 		$wordpress_order_status = $this->get_gateway_settings()['bitpay_checkout_order_process_confirmed_status'];
 		if ( WcGatewayBitpay::IGNORE_STATUS_VALUE === $wordpress_order_status ) {
 			$order->add_order_note(
-				$this->get_start_order_note( $invoice_id ) . ' has changed to Confirmed.  The order status has not been updated due to your settings.'
+				$this->get_start_order_note( $invoice_id ) . 'has changed to Confirmed. The order status has not been updated due to your settings.'
 			);
 			return;
 		}
 
-		$new_status = $this->get_wc_order_statuses()[ $wordpress_order_status ] ?? 'Processing';
+		$new_status = $this->get_wc_order_statuses()[ $wordpress_order_status ] ?? null;
 		if ( ! $new_status ) {
 			$new_status             = 'Processing';
 			$wordpress_order_status = 'wc-pending';
@@ -171,18 +197,24 @@ class BitPayIpnProcess {
 
 	private function process_completed( Invoice $bitpay_invoice, WC_Order $order ): void {
 		$this->validate_bitpay_status_in_available_statuses( $bitpay_invoice, array( 'complete' ) );
+		$wc_order_status = $order->get_status();
 
 		$invoice_id             = $bitpay_invoice->getId();
 		$wordpress_order_status = $this->get_gateway_settings()['bitpay_checkout_order_process_complete_status'];
 		if ( WcGatewayBitpay::IGNORE_STATUS_VALUE === $wordpress_order_status ) {
 			$order->add_order_note(
 				$this->get_start_order_note( $invoice_id )
-				. 'has changed to Complete.  The order status has not been updated due to your settings.'
+				. 'has changed to Complete. The order status has not been updated due to your settings.'
 			);
 			return;
 		}
 
-		$new_status = $this->get_wc_order_statuses()[ $wordpress_order_status ] ?? 'Processing';
+		if ( !$this->should_process_completed_action( $wc_order_status, $wordpress_order_status ) ) {
+			return;
+		}
+
+		$new_status = $this->get_wc_order_statuses()[ $wordpress_order_status ] ?? null;
+		$new_status = apply_filters( 'bitpay_checkout_order_process_complete_status', $new_status, $wordpress_order_status );
 		if ( ! $new_status ) {
 			$new_status             = 'Processing';
 			$wordpress_order_status = 'wc-pending';
@@ -214,7 +246,7 @@ class BitPayIpnProcess {
 		$invoice_id = $bitpay_invoice->getId();
 		$order->add_order_note(
 			$this->get_start_order_note( $invoice_id )
-			. 'has become invalid because of network congestion.  Order will automatically update when the status changes.'
+			. 'has become invalid because of network congestion. Order will automatically update when the status changes.'
 		);
 		$order->update_status( 'failed', __( 'BitPay payment invalid', 'woocommerce' ) );
 	}
@@ -273,5 +305,29 @@ class BitPayIpnProcess {
 
 		$new_status = $this->get_wc_order_statuses()[ $wordpress_order_status ] ?? 'processing';
 		$order->update_status( $new_status, __( 'BitPay payment processing', 'woocommerce' ) );
+	}
+
+	private function has_final_status( WC_Order $order ): bool {
+		return \in_array( $order->get_status(), self::FINAL_WC_ORDER_STATUSES, true );
+	}
+
+	/**
+	 * We don't want to change complete order to process.
+	 *
+	 * @param string      $wc_order_status WC order status.
+	 * @param string|null $wordpress_order_status_from_settings status to event from BitPay settings.
+	 * @return bool
+	 */
+	private function should_process_completed_action(string $wc_order_status, ?string $wordpress_order_status_from_settings ): bool
+	{
+		if ( $wc_order_status !== 'wc-completed' ) {
+			return true;
+		}
+
+		if ( $wordpress_order_status_from_settings === 'wc-pending' || $wordpress_order_status_from_settings === 'wc-processing' ) {
+			return false;
+		 }
+
+		return true;
 	}
 }